feat(setup): add internationalization
