+++ app/layout.tsx
@@ -1,22 +1,18 @@
 import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-import './globals.css';
-
-const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Hauser Metall',
+  description: 'Website der Firma Hauser Metall',
 };
 
-export default function RootLayout({
+export default function LocaleLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
+    <html lang={'de'}>
+      <body>{children}</body>
     </html>
   );
 }
