+++ globalConstants.ts
@@ -1,0 +1,5 @@
+/**
+ * List of all supported locales
+ * ATTENTION: The default locale is always the first one in the array.
+ * */
+export const locales: string[] = ['de', 'en'];
