+++ messages/de.json
@@ -1,0 +1,9 @@
+{
+  "Index": {
+    "title": "Hallo Welt!"
+  },
+  "Metadata": {
+    "title": "Hauser Metall",
+    "description": "Website der Firma Hauser Metall"
+  }
+}
