+++ messages/en.json
@@ -1,0 +1,9 @@
+{
+  "Index": {
+    "title": "Hello world!"
+  },
+  "Metadata": {
+    "title": "Hauser Metall",
+    "description": "Website of the Hauser Metall company"
+  }
+}
