+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "next": "14.2.1",
+    "next-intl": "3.11.3",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
